added new exception handlier

+++ src/main/java/com/shopping/service/impl/ShoppingServiceImpl.java
@@ -2,7 +2,8 @@ package com.shopping.service.impl;
 
 import com.shopping.dao.ShoppingDao;
 import com.shopping.entity.ShoppingEntity;
-import com.shopping.exception.ShoppingCustomException;
+import com.shopping.exception.ShoppingCustomBadRequestException;
+import com.shopping.exception.ShoppingCustomNotFoundException;
 import com.shopping.service.ShoppingService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -58,7 +59,7 @@ public class ShoppingServiceImpl implements ShoppingService {
         Optional<ShoppingEntity> existing = shoppingDao.findById(shopping.getShoppingId());
 
         if (existing.isPresent()) {
-            throw new ShoppingCustomException("DUPLICATE_SHOPPING_ID", "Shopping item with given id " +
+            throw new ShoppingCustomBadRequestException("DUPLICATE_SHOPPING_ID", "Shopping item with given id " +
                     shopping.getShoppingId() + " already exist");
         }
 
@@ -70,8 +71,12 @@ public class ShoppingServiceImpl implements ShoppingService {
 
         Optional<ShoppingEntity> single = shoppingDao.findById(shoppingId);
 
+        if (!isValidShoppingId(shoppingId)) {
+            throw new ShoppingCustomBadRequestException("INVALID_SHOPPING_ID", "Shopping item with ID " + shoppingId + " is invalid");
+        }
+
         if (single.isEmpty()) {
-            throw new ShoppingCustomException("NO_SHOPPING_ID_EXIST",
+            throw new ShoppingCustomNotFoundException("NO_SHOPPING_ID_EXIST",
                     "The shopping item with ID " + shoppingId + " is not present");
         }
 
@@ -79,10 +84,15 @@ public class ShoppingServiceImpl implements ShoppingService {
     }
 
     public List<ShoppingEntity> viewByProductName(String productName) {
+
+        if (productName == null || productName.isEmpty()) {
+            throw new ShoppingCustomBadRequestException("INVALID_PRODUCT_NAME", "Product name is mandatory");
+        }
+
         List<ShoppingEntity> singleProduct = shoppingDao.findByProductName(productName);
 
         if (singleProduct.isEmpty()) {
-            throw new ShoppingCustomException("NO_PRODUCT_NAME_EXIST",
+            throw new ShoppingCustomNotFoundException("NO_PRODUCT_NAME_EXIST",
                     "The product with name " + productName + " is not present");
         }
 
@@ -90,10 +100,15 @@ public class ShoppingServiceImpl implements ShoppingService {
     }
 
     public List<ShoppingEntity> viewByCustomerNameOrProductName(String customerName, String productName) {
+
+        if (customerName.isBlank() && customerName.isEmpty() && productName.isBlank() && productName.isEmpty()) {
+            throw new ShoppingCustomBadRequestException("INVALID_PRODUCT_AND_CUSTOMER_NAME", "Product or customer name is mandatory");
+        }
+
         List<ShoppingEntity> products = shoppingDao.findbyCustomerNameOrProductName(customerName, productName);
 
         if (products.isEmpty()) {
-            throw new ShoppingCustomException("NO_CUSTOMER_NAME_OR_PRODUCT_NAME_EXIST",
+            throw new ShoppingCustomNotFoundException("NO_CUSTOMER_NAME_OR_PRODUCT_NAME_EXIST",
                     "The products with name " + customerName + " or product name " + productName + " are not present");
         }
 
@@ -105,17 +120,22 @@ public class ShoppingServiceImpl implements ShoppingService {
         List<ShoppingEntity> every = shoppingDao.findAll();
 
         if (every.isEmpty()) {
-            throw new ShoppingCustomException("NO_SHOPPING_RECORDS", "Shopping list is empty");
+            throw new ShoppingCustomNotFoundException("NO_SHOPPING_RECORDS", "Shopping list is empty");
         }
         return every;
     }
 
     @Override
     public ShoppingEntity change(String shoppingId, ShoppingEntity shopping) {
+
+        if (!isValidShoppingId(shoppingId)) {
+            throw new ShoppingCustomBadRequestException("INVALID_SHOPPING_ID", "Shopping item with ID " + shoppingId + " is invalid");
+        }
+
         Optional<ShoppingEntity> exist = shoppingDao.findById(shoppingId);
 
         if (exist.isEmpty()) {
-            throw new ShoppingCustomException("NO_SHOPPING_RECORDS", "Shopping item of ID " + shoppingId + " is empty");
+            throw new ShoppingCustomNotFoundException("NO_SHOPPING_RECORDS", "Shopping item of ID " + shoppingId + " is empty");
         }
 
         ShoppingEntity shop = exist.get();
@@ -128,15 +148,15 @@ public class ShoppingServiceImpl implements ShoppingService {
         shop.setPurchaseModifyTime(dateToStringFormat(now()));
 
         if (shop.getBalanceAmount() < 0 || shop.getBuyingPrice() < shop.getSellingPrice()) {
-            throw new ShoppingCustomException("INVALID_BALANCE", "buying balance is low");
+            throw new ShoppingCustomBadRequestException("INVALID_BALANCE", "buying balance is low");
         }
 
         if (shopping.getProductName() == null || shopping.getProductName().isEmpty()) {
-            throw new ShoppingCustomException("PRODUCT_NAME_MANDATORY", "Product name is mandatory");
+            throw new ShoppingCustomBadRequestException("PRODUCT_NAME_MANDATORY", "Product name is mandatory");
         }
 
         if (!isValidEmail(shopping.getCustomerEmail())) {
-            throw new ShoppingCustomException("INVALID_EMAIL", "Entered email is invalid");
+            throw new ShoppingCustomBadRequestException("INVALID_EMAIL", "Entered email is invalid");
         }
 
         shoppingDao.save(shop);
@@ -147,38 +167,51 @@ public class ShoppingServiceImpl implements ShoppingService {
     @Transactional
     public ShoppingEntity changeByQuery(String shoppingId, ShoppingEntity shopping) {
 
+        if (shoppingId.isBlank() && shoppingId.isEmpty()) {
+            throw new ShoppingCustomBadRequestException("EMPTY_SHOPPING_ID", "Shopping ID should not be blank");
+        }
+
+        if (!isValidShoppingId(shoppingId)) {
+            throw new ShoppingCustomBadRequestException("INVALID_SHOPPING_ID", "Shopping item with ID " + shoppingId + " is invalid");
+        }
+
         Optional<ShoppingEntity> exist = shoppingDao.findById(shoppingId);
 
         if (exist.isEmpty()) {
-            throw new ShoppingCustomException("NO_SHOPPING_RECORDS", "Shopping item of ID " + shoppingId + " is empty");
+            throw new ShoppingCustomNotFoundException("NO_SHOPPING_RECORDS", "Shopping item of ID " + shoppingId + " is empty");
         }
 
         if (shopping.getProductName() == null || shopping.getProductName().isEmpty()) {
-            throw new ShoppingCustomException("PRODUCT_NAME_MANDATORY", "Product name is mandatory");
+            throw new ShoppingCustomBadRequestException("PRODUCT_NAME_MANDATORY", "Product name is mandatory");
         }
 
         if (!isValidEmail(shopping.getCustomerEmail())) {
-            throw new ShoppingCustomException("INVALID_EMAIL", "Entered email is invalid");
+            throw new ShoppingCustomBadRequestException("INVALID_EMAIL", "Entered email is invalid");
         }
 
         shopping.setCustomerName(extractName(shopping.getCustomerEmail()));
+        shopping.setBalanceAmount(shopping.getBuyingPrice() - shopping.getSellingPrice());
 
         shoppingDao.updateWithQuery(shoppingId, shopping.getProductName(), shopping.getCustomerName(), shopping.getCustomerEmail(),
-                shopping.getSellingPrice(), shopping.getBuyingPrice(), shopping.getBuyingPrice() - shopping.getSellingPrice(),
-                dateToStringFormat(now()));
+                shopping.getSellingPrice(), shopping.getBuyingPrice(), shopping.getBalanceAmount(), dateToStringFormat(now()));
 
         if (shopping.getBuyingPrice() < shopping.getSellingPrice()) {
-            throw new ShoppingCustomException("INVALID_BALANCE", "Buying price is less than selling price");
+            throw new ShoppingCustomBadRequestException("INVALID_BALANCE", "Buying price is less than selling price");
         }
 
         return shoppingDao.findById(shoppingId)
-                .orElseThrow(() -> new ShoppingCustomException("NO_SHOPPING_RECORDS", "Shopping item of ID " + shoppingId + " is empty"));
+                .orElseThrow(() -> new ShoppingCustomNotFoundException("NO_SHOPPING_RECORDS", "Shopping item of ID " + shoppingId + " is empty"));
     }
 
     @Override
     public void remove(String shoppingId) {
+
+        if (!isValidShoppingId(shoppingId)) {
+            throw new ShoppingCustomBadRequestException("INVALID_SHOPPING_ID", "Shopping item with ID " + shoppingId + " is invalid");
+        }
+
         if (!shoppingDao.existsById(shoppingId)) {
-            throw new ShoppingCustomException("SHOPPING_ITEM_NOT_FOUND",
+            throw new ShoppingCustomNotFoundException("SHOPPING_ITEM_NOT_FOUND",
                     "Shopping item with ID " + shoppingId + " not found");
         }
         shoppingDao.deleteById(shoppingId);
